Whenever we either discard a version due to its dependencies or perform conflict resolution, we return the last conflict that led to discarding them. In the main solver loop, we count those conflicts and pass their count to the user in `prioritize`.

In cargo, we use this information for prioritization, which speeds up resolution (`cargo run -r -- -m pub --with-solana --filter solana-archiver-lib -t 16` goes from 90s to 20s on my machine).

Configurations that are noticeably slower for the solana test case:
* All incompatibilities unit propagation
* Only the last root cause in unit propagation
* No incompatibilities from unit propagation
* No incompatibilities from `add_version`
* Only affect counts (without culprit counts)
* Backtracking with the same heuristic as https://github.com/astral-sh/uv/pull/9843 (backtracking once after affected hits 5)

In uv, we use this to re-prioritize and backtrack when a package decision accumulated to many conflicts. Since we have our own solver loop, we add the incompatibility to our own tracking instead.

Internally, we track more fields that we expose externally. This allows experimenting with different prioritization schemes without exposing potentially breaking changes in the API.

Built on https://github.com/pubgrub-rs/pubgrub/pull/291

## Benchmarks

Main:

```
        index commit hash: 82086e46740d7a9303216bfac093e7268a95121f
        index commit time: 2024-11-30T18:18:14Z
               index size: 32
          solana in index: 32
             Pub CPU time:  1215.49s ==  20.26min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    80.58s ==   1.34min
```

With https://github.com/pubgrub-rs/pubgrub/pull/291:

```
        index commit hash: 82086e46740d7a9303216bfac093e7268a95121f
        index commit time: 2024-11-30T18:18:14Z
               index size: 32
          solana in index: 32
             Pub CPU time:   467.73s ==   7.80min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    34.76s ==   0.58min
```

This PR:

```
        index commit hash: 82086e46740d7a9303216bfac093e7268a95121f
        index commit time: 2024-11-30T18:18:14Z
               index size: 32
          solana in index: 32
             Pub CPU time:   271.79s ==   4.53min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    20.17s ==   0.34min
```

Surprisingly, after adding conflicting-ness to the prioritization in any form, the should cancel count and the wall time lose their correlation, i've seen this effect with the other prioritization schemes i tried, too.

![Scatterplot](https://github.com/user-attachments/assets/bad49647-5127-49a6-826c-922cc4210d7f)

The remaining time seems to be spent mostly in semver(-pubgrub).

![Profile](https://github.com/user-attachments/assets/9276730c-4d63-45f2-ab98-a9cc44fd6e3f)