@Eh2406 2024-12-16
I do like the overall approach and particularly the code flexibility of this version. Unfortunately, I'm not able to reproduce the performance improvements.