@konstin 2024-12-17
What performance numbers do you get?