@x-hgg-x 2024-12-18
In my implementation I also get worse performance than #291: Pubgrub CPU time for `solana` crates goes from 350s to 520s.