@Eh2406 2024-12-18
Sorry I didn't get you numbers sooner.

https://github.com/Eh2406/pubgrub-crates-benchmark/commit/179f2895f9d5778e666fbe3fbd585fdc6672334e
```
> git checkout 179f2895f9d5778e666fbe3fbd585fdc6672334e

> cargo run -r -- -m pub --with-solana --filter solana-archiver-lib -t 16

   Compiling version-ranges v0.1.1 (https://github.com/Eh2406/pubgrub?branch=conflict_count#1044406b)
   Compiling cargo v0.86.0 (https://github.com/rust-lang/cargo?branch=master#876ea2ce)
   Compiling pubgrub v0.2.1 (https://github.com/Eh2406/pubgrub?branch=conflict_count#1044406b)
   Compiling semver-pubgrub v0.1.0 (https://github.com/pubgrub-rs/semver-pubgrub#0cfe3a1c)
   Compiling benchmark_from_crates v0.1.0 (/local/home/jfinkelm/PubGrub/pubgrub-crates-benchmark)
    Finished `release` profile [optimized + debuginfo] target(s) in 1m 09s
     Running `target/release/benchmark_from_crates -m pub --with-solana --filter solana-archiver-lib -t 16`
Running in mode Pub on 16 rayon threads.
!!!!!!!!!! Excluding Yanked !!!!!!!!!!
Start reading index
Done reading index in 26.4s
PubGrub: [Time: 36s, Rate: 0.8969/s, Remaining: 0s] 
!!!!!!!!!! Timings !!!!!!!!!!
        index commit hash: c03a36a091c45203caed5cbed4274a5f404a8c06
        index commit time: 2024-12-16T19:41:05Z
               index size: 32
          solana in index: 32
             Pub CPU time:   357.70s ==   5.96min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    35.68s ==   0.59min
```

https://github.com/Eh2406/pubgrub-crates-benchmark/commit/ec4830d8211003e5ec24c724fa1e9c7b7189288f
```
> git checkout ec4830d8211003e5ec24c724fa1e9c7b7189288f

> cargo run -r -- -m pub --with-solana --filter solana-archiver-lib -t 16


   Compiling version-ranges v0.1.1 (https://github.com/Eh2406/pubgrub?branch=conflict_count#65021252)
   Compiling pubgrub v0.2.1 (https://github.com/Eh2406/pubgrub?branch=conflict_count#65021252)
   Compiling semver-pubgrub v0.1.0 (https://github.com/pubgrub-rs/semver-pubgrub#0cfe3a1c)
   Compiling benchmark_from_crates v0.1.0 (/local/home/jfinkelm/PubGrub/pubgrub-crates-benchmark)
    Finished `release` profile [optimized + debuginfo] target(s) in 12.20s
     Running `target/release/benchmark_from_crates -m pub --with-solana --filter solana-archiver-lib -t 16`
Running in mode Pub on 16 rayon threads.
!!!!!!!!!! Excluding Yanked !!!!!!!!!!
Start reading index
Done reading index in 27.1s
PubGrub: [Time: 11s, Rate: 2.8661/s, Remaining: 0s] 
!!!!!!!!!! Timings !!!!!!!!!!
        index commit hash: c03a36a091c45203caed5cbed4274a5f404a8c06
        index commit time: 2024-12-16T19:41:05Z
               index size: 32
          solana in index: 32
             Pub CPU time:   129.76s ==   2.16min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    11.17s ==   0.19min

```

https://github.com/Eh2406/pubgrub-crates-benchmark/commit/65fea31a3fa63b0f02923240ea23f1c20bc1810d
```
> git checkout 65fea31a3fa63b0f02923240ea23f1c20bc1810d

> cargo run -r -- -m pub --with-solana --filter solana-archiver-lib -t 16

   Compiling version-ranges v0.1.1 (https://github.com/Eh2406/pubgrub?branch=konsti%2Fdev%2Fincompat-tracker#087bcdf5)
   Compiling pubgrub v0.2.1 (https://github.com/Eh2406/pubgrub?branch=konsti%2Fdev%2Fincompat-tracker#087bcdf5)
   Compiling semver-pubgrub v0.1.0 (https://github.com/pubgrub-rs/semver-pubgrub#0cfe3a1c)
   Compiling benchmark_from_crates v0.1.0 (/local/home/jfinkelm/PubGrub/pubgrub-crates-benchmark)
    Finished `release` profile [optimized + debuginfo] target(s) in 12.15s
     Running `target/release/benchmark_from_crates -m pub --with-solana --filter solana-archiver-lib -t 16`
Running in mode Pub on 16 rayon threads.
!!!!!!!!!! Excluding Yanked !!!!!!!!!!
Start reading index
Done reading index in 28.5s
PubGrub: [Time: 31s, Rate: 1.0488/s, Remaining: 0s] 
!!!!!!!!!! Timings !!!!!!!!!!
        index commit hash: c03a36a091c45203caed5cbed4274a5f404a8c06
        index commit time: 2024-12-16T19:41:05Z
               index size: 32
          solana in index: 32
             Pub CPU time:   459.65s ==   7.66min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    30.51s ==   0.51min
```