@Eh2406 2024-12-18
There were two big differences between my test of this branch and my branch. Obviously one of them was  how the metrics were collected. The other big change was the  removal of `discovery_order`,  I tried running my  branch but ignoring `discovery_order`:

https://github.com/Eh2406/pubgrub-crates-benchmark/commit/787462f05b0f046d1e85909bfda04070a3105ddf
```
   Compiling benchmark_from_crates v0.1.0 (/local/home/jfinkelm/PubGrub/pubgrub-crates-benchmark)
    Finished `release` profile [optimized + debuginfo] target(s) in 11.56s
     Running `target/release/benchmark_from_crates -m pub --with-solana --filter solana-archiver-lib -t 16`
Running in mode Pub on 16 rayon threads.
!!!!!!!!!! Excluding Yanked !!!!!!!!!!
Start reading index
Done reading index in 26.6s
!!!!!!!!!! Timings !!!!!!!!!!
        index commit hash: c03a36a091c45203caed5cbed4274a5f404a8c06
        index commit time: 2024-12-16T19:41:05Z
               index size: 32
          solana in index: 32
             Pub CPU time:   541.85s ==   9.03min
           Cargo CPU time: skipped
Cargo check lock CPU time: skipped
  Pub check lock CPU time: skipped
                Wall time:    46.21s ==   0.77min
```

Therefore, I feel comfortable merging this, as long as we follow it up fairly quickly by PR's adding `discovery_order`. Similarly any other differences in  what data is collected can be reevaluated in follow-up PR's.