@Eh2406 2024-12-18
In retrospect, the relevant number was "how much does this improve over dev"  not "how much does this improve over my PR". Sorry for the distraction. Over dev this is huge! `cargo run -r -- -m pub --with-solana --filter solana-archiver-lib -t 16` `dev: 24.08min` -> `this pr: 7.56min`.